+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -1,0 +1,54 @@
+package kitchenpos.order.application;
+
+import java.util.List;
+import kitchenpos.order.dto.OrderCreateRequest;
+import kitchenpos.order.dto.OrderUpdateRequest;
+import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.order.domain.OrderLineItemRepository;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderStatus;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+public class OrderService {
+    private final OrderRepository orderRepository;
+    private final OrderLineItemRepository orderLineItemRepository;
+    private final OrderValidator orderValidator;
+
+    public OrderService(OrderRepository orderRepository, OrderLineItemRepository orderLineItemRepository, OrderValidator orderValidator) {
+        this.orderRepository = orderRepository;
+        this.orderLineItemRepository = orderLineItemRepository;
+        this.orderValidator = orderValidator;
+    }
+
+    @Transactional
+    public Order create(final OrderCreateRequest request) {
+        Order order = Order.create(request.getOrderTableId(),
+                request.getOrderLineItems(),
+                orderValidator);
+
+        return orderRepository.save(order);
+    }
+
+    public List<Order> list() {
+        return orderRepository.findAll();
+    }
+
+    @Transactional
+    public Order changeOrderStatus(final Long orderId, final OrderUpdateRequest request) {
+        final Order savedOrder = orderRepository.findById(orderId)
+                .orElseThrow(IllegalArgumentException::new);
+
+        savedOrder.checkUpdatable();
+
+        final OrderStatus orderStatus = OrderStatus.valueOf(request.getOrderStatus());
+        savedOrder.setOrderStatus(orderStatus);
+
+        orderRepository.save(savedOrder);
+
+        savedOrder.setOrderLineItems(orderLineItemRepository.findAllByOrderId(orderId));
+
+        return savedOrder;
+    }
+}
