+++ src/main/java/kitchenpos/order/domain/OrderLineItemRepository.java
@@ -1,12 +1,10 @@
-package kitchenpos.dao;
-
-import kitchenpos.domain.OrderLineItem;
+package kitchenpos.order.domain;
 
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderLineItemDao extends JpaRepository<OrderLineItem, Long> {
+public interface OrderLineItemRepository extends JpaRepository<OrderLineItem, Long> {
     OrderLineItem save(OrderLineItem entity);
 
     Optional<OrderLineItem> findById(Long id);
