+++ src/main/java/kitchenpos/order/domain/OrderRepository.java
@@ -1,12 +1,10 @@
-package kitchenpos.dao;
-
-import kitchenpos.domain.Order;
+package kitchenpos.order.domain;
 
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderDao extends JpaRepository<Order, Long> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
     Order save(Order entity);
 
     Optional<Order> findById(Long id);
