+++ src/main/java/kitchenpos/order/dto/OrderCreateRequest.java
@@ -1,8 +1,8 @@
-package kitchenpos.application.dto;
+package kitchenpos.order.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItem;
 
 public class OrderCreateRequest {
     private Long orderTableId;
