+++ src/main/java/kitchenpos/order/dto/OrderUpdateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto;
+package kitchenpos.order.dto;
 
 public class OrderUpdateRequest {
     private String orderStatus;
