+++ src/test/java/kitchenpos/order/OrderStatusTest.java
@@ -1,8 +1,9 @@
-package kitchenpos.domain;
+package kitchenpos.order;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
+import kitchenpos.order.domain.OrderStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
