+++ src/test/java/kitchenpos/order/OrderTest.java
@@ -1,12 +1,12 @@
-package kitchenpos.domain;
+package kitchenpos.order;
 
-import static kitchenpos.support.Fixture.createOrderLineItem;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
-import java.util.Arrays;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
